# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kaspi Pay POS Automation API",
    "version": "1.0.0",
    "description": "REST API for integrating POS systems with Kaspi Pay API (SMS Authorization, Invoicing, QR Payments, Operation History, and Refunds)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Theater API Key used for authentication and permission gating"
      },
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Master admin secret token check (ADMIN_SECRET_KEY)"
      },
      "TokenSN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-SN",
        "description": "Session token SN obtained after OTP verification"
      },
      "VtokenSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Vtoken-Secret",
        "description": "Encrypted session token secret"
      },
      "ProfileId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Profile-Id",
        "description": "Organization profile ID (optional, sent after auth)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Description of the error"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/orders/invoice": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create Custom Order Invoice",
        "description": "Generates a remote invoice on Kaspi linked to a unique customer order and upserts the event context inside the local database.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId", "eventId", "eventName", "amount", "customerPhone"],
                "properties": {
                  "orderId": {
                    "type": "string",
                    "description": "Unique order identifier from the ticketing system",
                    "example": "SM-2026-00481"
                  },
                  "eventId": {
                    "type": "string",
                    "description": "Unique event identifier",
                    "example": "moldanazar-live"
                  },
                  "eventName": {
                    "type": "string",
                    "description": "Display name of the event",
                    "example": "Moldanazar — Live Concert"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Invoice payment amount in Tenge",
                    "example": 8000
                  },
                  "customerPhone": {
                    "type": "string",
                    "description": "Customer mobile phone number in format 7XXXXXXXXXX",
                    "example": "77077008371"
                  },
                  "customerName": {
                    "type": "string",
                    "description": "Customer full name (optional)",
                    "example": "Orken"
                  },
                  "customerEmail": {
                    "type": "string",
                    "description": "Customer email address (optional)",
                    "example": "orken@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "integer",
                          "example": 123456
                        },
                        "Status": {
                          "type": "string",
                          "example": "RemotePaymentCreated"
                        },
                        "Amount": {
                          "type": "number",
                          "example": 8000
                        },
                        "ClientMobile": {
                          "type": "string",
                          "example": "77077008371"
                        },
                        "ReceiptUrl": {
                          "type": "string",
                          "example": "https://..."
                        },
                        "OrderNumber": {
                          "type": "string",
                          "example": "ORDER-123"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server or Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Manually Override Order Status",
        "description": "Allows administrators to manually set the payment status of an order (e.g. Processed, Cancelled, Refunded) and automatically halts any active background polling for this transaction.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique database auto-incrementing ID or the custom POS order number",
            "schema": {
              "type": "string"
            },
            "example": "SM-2026-00481"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "The new payment status. Note: 'CancelledByUser' is client-side QR payment cancellation, 'RemotePaymentCanceled' is client-side remote invoice rejection, and 'Cancelled' is manual administrator cancellation.",
                    "enum": [
                      "pending",
                      "Processed",
                      "Wait",
                      "QrTokenCreated",
                      "RemotePaymentCreated",
                      "RemotePaymentCanceled",
                      "CancelledByUser",
                      "Cancelled",
                      "Expired",
                      "Error",
                      "Failed",
                      "Refunded"
                    ],
                    "example": "Processed"
                  },
                  "status_desc": {
                    "type": "string",
                    "description": "Optional description or reason explaining the manual override",
                    "example": "Manually marked as paid by administrator"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status manually overridden successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Order status manually updated to Processed."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing status parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found or belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}